Fix the Max NPCs per Platform setting
Previously the only thing it limited was the toolgun UI which isn't
really much of a protection.

This also limits it to integer values on the admin panel for obvious
reasons

+++ lua/autorun/client/npcspawner2_cl.lua
@@ -94,7 +94,7 @@ local function adminOptions(panel)
 			Label = lang("maxinplay"),
 			Command = cvar("maxinplay"),
 			Help = true,
-			Type = "Float",
+			Type = "Int",
 			Min = "1",
 			Max = "50",
 		}
